# backend/handlers/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PAIMOS API",
    "description": "PAIMOS — Professional & Personal AI Project OS. Stable HTTP API used by the SPA, the paimos CLI, and any custom integration. Authentication is either a session cookie (browser) or a `paimos_*` bearer API key (machine-to-machine).",
    "version": "1.0.0",
    "license": { "name": "AGPL-3.0-only" }
  },
  "servers": [
    { "url": "/", "description": "Same-origin (recommended)" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Issued by POST /api/auth/login. Browser SPA. CSRF-protected via X-CSRF-Token header for mutations (PAI-113)."
      },
      "bearerAPIKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "paimos_*",
        "description": "Long-lived API key, prefix configurable via BRAND_API_KEY_PREFIX. Created via /api/auth/api-keys."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Human-readable error message" }
        }
      },
      "KnowledgeEntry": {
        "type": "object",
        "description": "PAI-394 unified knowledge entry shape. `type` is the SQL discriminator (snake_case). Returned by every read on /api/projects/{id}/knowledge.",
        "required": ["id", "project_id", "type", "slug", "title", "status"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "project_id": { "type": "integer", "format": "int64" },
          "type": { "type": "string", "enum": ["memory", "runbook", "guideline", "external_system", "related_project"] },
          "slug": { "type": "string", "pattern": "^[a-z][a-z0-9_-]*$", "maxLength": 64 },
          "title": { "type": "string" },
          "body": { "type": "string", "description": "Markdown body" },
          "status": { "type": "string", "description": "Issue status enum value (typically backlog / proposed / cancelled)" },
          "metadata": { "type": "object", "additionalProperties": true, "description": "Per-type tail fields (schemaless on the server; per-Module validation runs on write)" },
          "created_at": { "type": "string" },
          "updated_at": { "type": "string" },
          "reference_count": { "type": "integer", "description": "PAI-347 decay counter (memory only; 0 for other types)" },
          "last_referenced_at": { "type": "string" }
        }
      },
      "KnowledgeEntryInput": {
        "type": "object",
        "description": "Write shape for POST / PUT on the unified knowledge surface. `type` is required on POST (either kebab URL segment or snake_case discriminator); on PUT, the URL pins the type and a body type must match.",
        "required": ["slug", "title"],
        "properties": {
          "type": { "type": "string", "description": "Knowledge type — `memory`, `runbook`, `guideline`, `external-system`/`external_system`, `related-project`/`related_project`" },
          "slug": { "type": "string", "pattern": "^[a-z][a-z0-9_-]*$", "maxLength": 64 },
          "title": { "type": "string" },
          "body": { "type": "string" },
          "status": { "type": "string", "description": "Optional; defaults to the type's DefaultStatus" },
          "metadata": { "type": "object", "additionalProperties": true }
        }
      },
...
      "Issue": {
        "type": "object",
        "required": ["id", "project_id", "title", "type", "status"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "project_id": { "type": "integer", "format": "int64", "nullable": true },
          "title": { "type": "string" },
          "type": { "type": "string", "enum": ["epic", "cost_unit", "release", "sprint", "ticket", "task"] },
          "status": { "type": "string", "enum": ["new", "backlog", "in-progress", "qa", "done", "delivered", "accepted", "invoiced", "cancelled"] },
          "priority": { "type": "string", "enum": ["low", "medium", "high"] },
          "description": { "type": "string" },
          "acceptance_criteria": { "type": "string" },
          "notes": { "type": "string" },
          "parent_id": { "type": "integer", "format": "int64", "nullable": true },
          "assignee_id": { "type": "integer", "format": "int64", "nullable": true },
          "cost_unit": { "type": "string" },
          "release": { "type": "string" },
          "issue_number": { "type": "integer" },
          "key": { "type": "string", "description": "Human-friendly identifier {project_key}-{issue_number}" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "responses": {
...
      "NotFound": {
        "description": "Resource not found, or hidden from caller for privacy",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [
    { "sessionCookie": [] },
    { "bearerAPIKey": [] }
  ],
  "paths": {
...
    "/api/issues/{id}": {
      "get": {
        "tags": ["issues"],
        "summary": "Get issue (numeric id or {project_key}-{n})",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Issue", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Issue" } } } }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "put": {
        "tags": ["issues"],
        "summary": "Update issue",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Updated" } }
      },
      "patch": {
        "tags": ["issues"],
        "summary": "Update issue",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Updated" } }
      }
    },
...
    "/api/projects/{id}/knowledge": {
      "get": {
        "tags": ["knowledge"],
        "summary": "List knowledge entries (all types or filtered)",
        "description": "PAI-394 unified surface. Without `type`, returns every non-trashed knowledge entry across all five registered types. With `type=<segment>`, narrows to that type. Segments are kebab-singular: memory, runbook, guideline, external-system, related-project.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "type", "in": "query", "required": false, "schema": { "type": "string", "enum": ["memory", "runbook", "guideline", "external-system", "related-project"] } }
        ],
        "responses": { "200": { "description": "Array of knowledge entries", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/KnowledgeEntry" } } } } } }
      },
      "post": {
        "tags": ["knowledge"],
        "summary": "Create a knowledge entry",
        "description": "Type travels either in the body (`type` field, kebab segment or snake_case discriminator) or as a `?type=` query parameter — body wins on conflict.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "type", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Convenience: matches body.type when both are present" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeEntryInput" } } }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeEntry" } } } },
          "400": { "description": "Invalid type, slug, or per-type metadata" },
          "409": { "description": "Slug already exists for this type in this project" }
        }
      }
    },
    "/api/projects/{id}/knowledge/{type}/{slug}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
        { "name": "type", "in": "path", "required": true, "schema": { "type": "string", "enum": ["memory", "runbook", "guideline", "external-system", "related-project"] } },
        { "name": "slug", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[a-z][a-z0-9_-]*$" } }
      ],
      "get": {
        "tags": ["knowledge"],
        "summary": "Fetch a single knowledge entry",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeEntry" } } } },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "tags": ["knowledge"],
        "summary": "Replace a knowledge entry (rename via body.slug)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeEntryInput" } } }
        },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeEntry" } } } },
          "404": { "description": "Not found" },
          "409": { "description": "Slug rename collides with an existing entry of the same type" }
        }
      },
      "delete": {
        "tags": ["knowledge"],
        "summary": "Soft-delete (Trash) a knowledge entry",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  }
}